total function done

+++ cpp/grains/grains.h
@@ -2,8 +2,8 @@
 #define GRAINS_H
 
 namespace grains {
-  unsigned long square(const unsigned long kNumber);
-  unsigned long total();
+  unsigned long long square(const unsigned long long kNumber);
+  unsigned long long total();
 }  // namespace grains
 
 #endif // GRAINS_H
